refactor spaceship builder
